+++ tests/integration/test_config_handling.py
@@ -14,11 +14,13 @@
 
 import sys
 import warnings
+from io import StringIO
 
 from delayed_assert import expect, assert_expectations
 from unittest import mock
 
 from examples.test_rp_logging import LOG_MESSAGE
+from pytest_reportportal.config import OUTPUT_TYPES
 from tests import REPORT_PORTAL_SERVICE
 from tests.helpers import utils
 
@@ -282,4 +284,90 @@ def test_retries(mock_client_init):
         constructor_args = mock_client_init.call_args_list[0][1]
         expect(constructor_args['retries'] == retries)
         expect(len(filter_agent_calls(w)) == 1)
+    assert_expectations()
+
+
+@mock.patch(REPORT_PORTAL_SERVICE)
+def test_launch_uuid_print(mock_client_init):
+    print_uuid = True
+    variables = utils.DEFAULT_VARIABLES.copy()
+    variables.update({'rp_launch_uuid_print': str(print_uuid)}.items())
+
+    str_io = StringIO()
+    stdout = sys.stdout
+    try:
+        OUTPUT_TYPES['stdout'] = str_io
+        result = utils.run_pytest_tests(['examples/test_rp_logging.py'],
+                                        variables=variables)
+    finally:
+        OUTPUT_TYPES['stdout'] = stdout
+
+    assert int(result) == 0, 'Exit code should be 0 (no errors)'
+    expect(mock_client_init.call_count == 1)
+
+    expect('Report Portal Launch UUID:' in str_io.getvalue())
+    assert_expectations()
+
+
+@mock.patch(REPORT_PORTAL_SERVICE)
+def test_launch_uuid_print_stderr(mock_client_init):
+    print_uuid = True
+    variables = utils.DEFAULT_VARIABLES.copy()
+    variables.update({'rp_launch_uuid_print': str(print_uuid), 'rp_launch_uuid_print_output': 'stderr'}.items())
+
+    str_io = StringIO()
+    stderr = sys.stderr
+    try:
+        OUTPUT_TYPES['stderr'] = str_io
+        result = utils.run_pytest_tests(['examples/test_rp_logging.py'],
+                                        variables=variables)
+    finally:
+        OUTPUT_TYPES['stderr'] = stderr
+
+    assert int(result) == 0, 'Exit code should be 0 (no errors)'
+    expect(mock_client_init.call_count == 1)
+
+    expect('Report Portal Launch UUID:' in str_io.getvalue())
+    assert_expectations()
+
+
+@mock.patch(REPORT_PORTAL_SERVICE)
+def test_launch_uuid_print_invalid_output(mock_client_init):
+    print_uuid = True
+    variables = utils.DEFAULT_VARIABLES.copy()
+    variables.update({'rp_launch_uuid_print': str(print_uuid), 'rp_launch_uuid_print_output': 'something'}.items())
+
+    str_io = StringIO()
+    stdout = sys.stdout
+    try:
+        OUTPUT_TYPES['stdout'] = str_io
+        result = utils.run_pytest_tests(['examples/test_rp_logging.py'],
+                                        variables=variables)
+    finally:
+        OUTPUT_TYPES['stdout'] = stdout
+
+    assert int(result) == 0, 'Exit code should be 0 (no errors)'
+    expect(mock_client_init.call_count == 1)
+
+    expect('Report Portal Launch UUID:' in str_io.getvalue())
+    assert_expectations()
+
+
+@mock.patch(REPORT_PORTAL_SERVICE)
+def test_no_launch_uuid_print(mock_client_init):
+    variables = utils.DEFAULT_VARIABLES.copy()
+
+    str_io = StringIO()
+    stdout = sys.stdout
+    try:
+        OUTPUT_TYPES['stdout'] = str_io
+        result = utils.run_pytest_tests(['examples/test_rp_logging.py'],
+                                        variables=variables)
+    finally:
+        OUTPUT_TYPES['stdout'] = stdout
+
+    assert int(result) == 0, 'Exit code should be 0 (no errors)'
+    expect(mock_client_init.call_count == 1)
+
+    expect('Report Portal Launch UUID:' not in str_io.getvalue())
     assert_expectations()